We create a class called GoogleMaps.
First we store the center and date being passed in from our app to the GoogleMap instance class.

Then we invoke createElement() which uses our parent class BaseElement. Its mission is to call getElementString() and then instantiate the DOM object being represented by the provided string.

We override in GoogleMaps the parent's getElementString() method to create a div with id map.

Then, we continue createElement execution. The fact is that we need to work with the div being created, with the DOM element, so then we set a timeout to wait until the next CPU cycle.

Then we put a global variable into window's scope to call to google api and create a Map, which will be rendered in the div, with id map.

In addition we get all the vehiccles passed in and iterate through them. We unpack the latitude and longitude of each vehicle, knowing that them are separated by a space.

Then we store in a global window variable the latitude and longitude which we store.

After that we create a Marker, on the position of the vehicle and on the map.

Finally we show the marker on top of the map.

Nice, isn't it?