We convert our App in a ES6 class. It extends from ApplicationBase because the idea is to put all shared behaviour and properties between our apps in ApplicationBase.
Then App's aim is to bootstrap the application with a title, and load all the JSON data needed.

After that we make our App object accesible to all web pages. It would be needed to allow routing.

Finally we implement a method show(element) in ApplicationBase because we want to append the components which will make our application on the HTML.

;=)