In App.js we add the routes, which means we pass in the title, the page to go, and if it is the default loaded page.
Then we implement the addRoute() method in the parent class: ApplicationBase. It will be a shared method among all apps.

We accept the id which is the string which is shown, the pageObject to go to, and if it is the defaultRoute to load on app init.

Then we pass in the id to show the strings for links, and the pageObject to link.

Adter that we associate in a map, an object with pairs key value, the id and the pageobject.

;=)!!!