We create a HomePage which is a div with its children centered.
In addition we add a large Image to the HomePage.

Then we style and add a button for cars and other for drones.

In button js we add the capability to stile it using a class string. Then we inject it into the HTML markup.

In applicationBase we store the dafault rout's name to be loaded when app inits.
It is interesting to note that when the defaultRoute is parsed when showing the app, we find the class which is defined by material designed, call .page-content to append our HTML HomePage markup. First we empty whatever markup the previous page had, and then we see the routeMap and load the desired content.

;=) ;=) ;=)

Finally is in app.js where we associate the title Home with the HomePage class and set it to the defaultRoute variable ;=) ;=) ;) ;=)