We add a new abstract class called Page with the idea to put all the shared components between pages like breadcrumbs or footers.
Then we create the CarsPage whichs is a div with a heading and after we create and append the table with the JSON cars' data.

After that, we remove the href inside the links into the titleBar to take control of it with javascript and avoid default browser behaviour.

We modify applicationBase to define how to find all titleBar links, and associate them an onclick handler, to load the route of the title clicked.

We adapt the HomePage to inherit from Page

Intereting!